requirements.txt
# Core ML packages
torch>=2.1.0
torchvision>=0.17.0
transformers>=4.36.0
sentencepiece==0.1.99

# MLX packages
mlx>=0.0.8

# RAG and NLP packages
rank_bm25>=0.2.2
nltk>=3.8.1
dspy-ai>=2.3.4

# Ollama integration
ollama>=0.1.6

# Development tools
jupyter>=1.0.0
ipykernel>=6.29.0

# Additional dependencies
numpy>=1.24.0
pandas>=2.2.0
scikit-learn>=1.3.0
dataclasses>=0.6
typing>=3.7.4.3 